Refactor navigation renderer specs
