+++ src/heart/navigation/composed_renderer.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import Protocol, TypeVar
 
 import pygame
 
@@ -9,13 +8,8 @@ from heart.device import Orientation
 from heart.peripheral.core.manager import PeripheralManager
 from heart.renderers import StatefulBaseRenderer
 
-RendererT = TypeVar("RendererT", bound=StatefulBaseRenderer)
-RendererSpec = StatefulBaseRenderer | type[StatefulBaseRenderer]
-
-
-class RendererResolver(Protocol):
-    def resolve(self, dependency: type[RendererT]) -> RendererT:
-        """Resolve renderer instances from the shared container."""
+from .renderer_specs import (RendererResolver, RendererSpec,
+                             resolve_renderer_spec)
 
 
 @dataclass
@@ -34,7 +28,10 @@ class ComposedRenderer(StatefulBaseRenderer[ComposedRendererState]):
     ) -> None:
         super().__init__()
         self._renderer_resolver = renderer_resolver
-        self.renderers = [self._resolve_renderer_spec(renderer) for renderer in renderers]
+        self.renderers = [
+            resolve_renderer_spec(renderer, renderer_resolver, "ComposedRenderer")
+            for renderer in renderers
+        ]
 
     def _real_get_renderers(self) -> list[StatefulBaseRenderer]:
         result: list[StatefulBaseRenderer] = []
@@ -65,7 +62,8 @@ class ComposedRenderer(StatefulBaseRenderer[ComposedRendererState]):
 
     def add_renderer(self, *renderers: RendererSpec) -> None:
         resolved_renderers = [
-            self._resolve_renderer_spec(renderer) for renderer in renderers
+            resolve_renderer_spec(renderer, self._renderer_resolver, "ComposedRenderer")
+            for renderer in renderers
         ]
         self.renderers.extend(resolved_renderers)
         if self.is_initialized():
@@ -94,17 +92,6 @@ class ComposedRenderer(StatefulBaseRenderer[ComposedRendererState]):
         self, renderer: type[StatefulBaseRenderer]
     ) -> None:
         self.add_renderer(renderer)
-
-    def _resolve_renderer_spec(self, renderer: RendererSpec) -> StatefulBaseRenderer:
-        if isinstance(renderer, type):
-            if not issubclass(renderer, StatefulBaseRenderer):
-                raise TypeError(
-                    "ComposedRenderer requires StatefulBaseRenderer subclasses"
-                )
-            if self._renderer_resolver is None:
-                raise ValueError("ComposedRenderer requires a renderer resolver")
-            return self._renderer_resolver.resolve(renderer)
-        return renderer
 
     def real_process(
         self,
