+++ src/heart/navigation/multi_scene.py
@@ -9,7 +9,8 @@ from heart.peripheral.core.manager import PeripheralManager
 from heart.peripheral.switch import SwitchState
 from heart.renderers import StatefulBaseRenderer
 
-from .composed_renderer import RendererResolver, RendererSpec
+from .renderer_specs import (RendererResolver, RendererSpec,
+                             resolve_renderer_spec)
 
 
 @dataclass
@@ -26,7 +27,10 @@ class MultiScene(StatefulBaseRenderer[MultiSceneState]):
     ) -> None:
         super().__init__()
         self._renderer_resolver = renderer_resolver
-        self.scenes = [self._resolve_renderer_spec(scene) for scene in scenes]
+        self.scenes = [
+            resolve_renderer_spec(scene, renderer_resolver, "MultiScene")
+            for scene in scenes
+        ]
 
     def get_renderers(self) -> list[StatefulBaseRenderer]:
         index = self._active_scene_index()
@@ -65,12 +69,3 @@ class MultiScene(StatefulBaseRenderer[MultiSceneState]):
     def _active_scene_index(self) -> int:
         offset = self.state.offset_of_button_value or 0
         return (self.state.current_button_value - offset) % len(self.scenes)
-
-    def _resolve_renderer_spec(self, renderer: RendererSpec) -> StatefulBaseRenderer:
-        if isinstance(renderer, type):
-            if not issubclass(renderer, StatefulBaseRenderer):
-                raise TypeError("MultiScene requires StatefulBaseRenderer subclasses")
-            if self._renderer_resolver is None:
-                raise ValueError("MultiScene requires a renderer resolver")
-            return self._renderer_resolver.resolve(renderer)
-        return renderer
